server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pyyaml==6.0.1
python-dotenv==1.0.0
pydantic==2.5.3
requests==2.31.0
